# Optimize `brew update` by batching `git cat-file` operations