--- a/Library/Homebrew/test/cmd/update-report_spec.rb
+++ b/Library/Homebrew/test/cmd/update-report_spec.rb
@@ -27,8 +27,9 @@ describe "brew update-report" do
     let(:hub) { ReporterHub.new }
 
     def perform_update(fixture_name = "")
-      allow(Formulary).to receive(:factory).and_return(double(pkg_version: "1.0"))
-      allow(FormulaeAtRevision).to receive(:new).and_return(double("[]": double(pkg_version: "2.0")))
+      formula = double(pkg_version: "1.0")
+      allow(Formulary).to receive(:factory).and_return(formula)
+      allow(FormulaVersions).to receive(:formulae_at_repo_revision).and_return(formula => double(pkg_version: "2.0"))
 
       diff = YAML.load_file("#{TEST_FIXTURE_DIR}/updater_fixture.yaml")[fixture_name]
       allow(reporter).to receive(:diff).and_return(diff || "")