@boblail 2022-05-04
@MikeMcQuaid, this approach makes lines like this in the specs inert:
```ruby
allow(FormulaVersions).to receive(:new).and_return(double(formula_at_revision: "2.0"))
```

If you prefer this approach to #13243, what'd be your inclination for architecture?

My first thought was to extract a class like this — that could be stubbed in a similar way:
```ruby
class FormulaeAtRevision
  def initialize(formulae, revision)
    # ...
  end

  # accepts the current version of a formula, return its version at `@revision`
  def version_of(formula)
    # ... 
  end
end
```