@boblail 2022-05-05
@MikeMcQuaid, would you be receptive to a PR that kept `git cat-file --batch` as a strategy but implemented it in `FormulaVersions` instead of `FormulaeAtRevision`?