@MikeMcQuaid 2022-05-05
@boblail Potentially: depends on how much the complexity increases. I'm not really convinced that the solution here is in `git cat-file --batch` usage but more: