@MikeMcQuaid 2022-05-05
- checking fewer revisions at all
- checking fewer formulae
- not checking formulae at all in some cases
- caching versions