@boblail 2022-05-05
@MikeMcQuaid, what do you think of this refactor?

It
- discards `FormulaeAtRevision`
- extracts `GitRepositoryExtension#git_cat_files`
   (which, I think, could also optimize [formula-auditor](https://github.com/Homebrew/brew/blob/master/Library/Homebrew/formula_auditor.rb#L719-L721) — instead of catting many formulae at one revision, we're catting one formula at many revisions, but I don't want to explode the scope of this PR 😬)
- implements `file_contents_at_revision` in terms of `repository.git_cat_file` and `formula_at_revision` in terms of `FormulaVersions.formula_from_contents` so
    1. there's not a duplication of logic between the one-off and bulk approaches for fetching previous versions of formulae
    2. the new strategy is better covered by other specs that were leveraging `FormulaVersions#formula_at_revision`