@boblail 2022-05-05
> What's the performance benefit here for the typical and pathological cases?

@MikeMcQuaid, would you consider running `brew update` once a week typical and once every 3 months pathological?

At Square, we nudge engineers to do this monthly and we regularly see outliers who haven't done it in several months.

If so, this branch makes `brew update`
- **3.7×** faster for the typical use-case (22.6s → 6.1s)
- **9.3×** faster for the pathological use-case (135s → 14.5s)