@MikeMcQuaid 2022-05-15
@boblail Yeh, that's a decent speedup, thanks. I'm going to want to iterate a lot on the implementation here so I'd rather if we can land something like https://github.com/Homebrew/brew/pull/13243#issuecomment-1118475144 to simplify the need for this instead.