@MikeMcQuaid 2022-05-18
I think https://github.com/Homebrew/brew/pull/13299 may remove the need for this but interested in thoughts.