@MikeMcQuaid 2022-06-09
Passing on this given recent `brew update` changes. Feel free to reopen if still needed/desired.